requirements.txt
django~=3.1.7
django-allauth
djangorestframework~=3.12.2